+++ app/lib/ui/utilities/app_assets.dart
@@ -27,15 +27,19 @@ class _Images {
   final String ilComponentsButtonDark = 'assets/il_components_button_dark.svg';
 
   final String ilComponentsDivider = 'assets/il_components_divider.svg';
-  final String ilComponentsDividerDark = 'assets/il_components_divider_dark.svg';
+  final String ilComponentsDividerDark =
+      'assets/il_components_divider_dark.svg';
 
   final String ilcomponentsCheckbox = 'assets/il_components_checkbox.svg';
-  final String ilcomponentsCheckboxDark = 'assets/il_components_checkbox_dark.svg';
+  final String ilcomponentsCheckboxDark =
+      'assets/il_components_checkbox_dark.svg';
 
-  final String ilComponentsRadioButton = 'assets/il_components_radio_button.svg';
-  final String ilComponentsRadioButtonDark = 'assets/il_components_radio_button_dark.svg';
+  final String ilComponentsRadioButton =
+      'assets/il_components_radio_button.svg';
+  final String ilComponentsRadioButtonDark =
+      'assets/il_components_radio_button_dark.svg';
 
-  ///Tokens
+  /// Tokens
   final String ilTokensColor = 'assets/il_tokens_color.svg';
   final String ilTokensColorDark = 'assets/il_tokens_color_dark.svg';
 
@@ -47,6 +51,9 @@ class _Images {
 
   final String ilUnion = 'assets/il_union.svg';
   final String ilUnionDark = 'assets/il_union_dark.svg';
+
+  final String ilTypography = 'assets/il_tokens_typography.svg';
+  final String ilTypographyDark = 'assets/il_tokens_typography_dark.svg';
 }
 
 class _Icons {